refactor: move ProtectedRoute.tsx from routes to shared components
